Fix outstanding issues with the lamp demo.
Save the current state in redis and send it on connect.
Redis doesn't give the last message to new subscribers,
so this is needed to initialize the button.

Make things more self-referential and clean up logging.

Use an <input type=button> instead of a <button> because
it's simpler to use the value for the display text that way.